% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence_corr_method.R
\name{correlation}
\alias{correlation}
\title{Correlation}
\usage{
correlation(data)
}
\arguments{
\item{data}{Input matrix with proper column names.}
}
\value{
Correlation matrix with column names inherited from the data
}
\description{
Correlation
}
