% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dependency.explorer}
\alias{dependency.explorer}
\title{eXSD Dependency Explorer}
\usage{
dependency.explorer(
  datasets,
  labels,
  fs = 1,
  port = NULL,
  launch.browser = TRUE
)
}
\arguments{
\item{datasets}{Named list of matrices
\code{list( code1=matrix1, code2=matrix2, code3=matrix3 )}}

\item{labels}{Named list of labels to show in the UI
\code{list( code1="label 1", code2="label 2", code3="label 3" )}}

\item{port}{Port when the RShiny app will open}

\item{launch.browser}{Open browser. Default: TRUE.}

\item{sampling_frequency}{Sampling rate at the datasets were collected.}
}
\description{
\figure{logo.png}{options: width="100px" alt="eXSD Dependency Explorer"}

Start the main visual interface of the dependency explorer.
}
