% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_csv_utils.R
\name{load_csv_datasets}
\alias{load_csv_datasets}
\title{Load a set of CSV files}
\usage{
load_csv_datasets(path, filenames, length = NULL)
}
\arguments{
\item{path}{Directory where the files are located}

\item{filenames}{List with filenames to load.
\code{list( code1="/path1", code2="/path2", code3="/path3" )}}

\item{length}{Maximum number of points per file}
}
\value{
List with file contents ordered according to the names:

\code{list( code1=matrix1, code2=matrix1, code3=matrix1 )}
}
\description{
Load a set of CSV files
}
