% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_csv_utils.R
\name{load_partitioned_csv_dataset}
\alias{load_partitioned_csv_dataset}
\title{Load a CSV file (split in segments)}
\usage{
load_partitioned_csv_dataset(
  path,
  range = list(start = 1, end = NULL, step = 200)
)
}
\arguments{
\item{path}{.   .}

\item{range}{Sliding window settings: \code{list(start, stop, step)}}
}
\value{
List with the sliding windows

\code{list( matrix_segment1, matrix_segment2, matrix_segment3 )}
}
\description{
Load a CSV file (split in segments)
}
