% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg_simulate_eeg.R
\name{simulate_EEG}
\alias{simulate_EEG}
\title{Simulate EEG}
\usage{
simulate_EEG(
  N,
  fs = 200,
  f = c(0, 2, 4, 10, 20, 34, 45),
  tau = c(1, 1, 1, 1, 1, 1, 1),
  sigma = c(3, 1, 1, 1, 1, 1, 1),
  electrical_line = NULL
)
}
\description{
Simulate EEG
}
\examples{
\dontrun{
  y <- simulate_EEG(1000) # Sample 1000 points at 200Hz
  spectrum(simulate_EEG(1000, electrical_line=50,), log="no")
}
}
