% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg_simulate_sparse_contemporaneous_mixture.R
\name{simulate_sparse_contemporaneous_mixture}
\alias{simulate_sparse_contemporaneous_mixture}
\title{\donotrun{
mixture_matrix <- rbind(
c(2, 1),
c(0, 0.5)
)
Y <- simulate_sparse_contemporaneous_mixture(N, mixture_matrix=mixture_matrix, fs=fs, f=c(2, 10), tau=c(5, 5), sigma=c(2, 2), sigma_epsilon=c(1, 1))
plot(Y\link{1,}, type="l")
lines(Y\link{2,})
}}
\usage{
simulate_sparse_contemporaneous_mixture(
  N,
  mixture_matrix,
  fs,
  lag_matrix = NULL,
  f = c(2, 10),
  tau = c(1, 1),
  sigma = c(2, 2),
  sigma_epsilon = c(1, 1)
)
}
\description{
\donotrun{
mixture_matrix <- rbind(
c(2, 1),
c(0, 0.5)
)
Y <- simulate_sparse_contemporaneous_mixture(N, mixture_matrix=mixture_matrix, fs=fs, f=c(2, 10), tau=c(5, 5), sigma=c(2, 2), sigma_epsilon=c(1, 1))
plot(Y\link{1,}, type="l")
lines(Y\link{2,})
}
}
